+++ CHANGELOG.md
@@ -5,11 +5,11 @@ _unreleased_
 **Breaking Changes**
 - Character for negative scope tests is now `!` (before: `-`)
 - Failure exit code is now `1` (before: `-1`)
-- Remove support for multiple assertions in one line, e.g. `^^  ^^^ source.xy`
+- Remove (the undocumented) support for multiple assertions in one line, e.g. `^^  ^^^ source.xy`
   - Workaround: split into multiple lines
 
 **Other Changes**
-- readme: Improve syntax explanation
+- readme: Clarify testing syntax
 - unit/parser: Complete code overhaul and refactor
 - Replace mocha/chai with bun:test